backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
cryptography
django-environ
psycopg2
django-cors-headers
django-filter

gunicorn
django-heroku